user delete

+++ backend/src/main/java/com/chatapp/backend/repository/UserRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User,Integer>{
 
-	void deleteById(long id);
+	void deleteById(Integer id);
     
 }
